Second Commit (modif loop)

+++ JOBSHEET 3/LATIHAN 1/mahasiswa04.java
@@ -2,6 +2,6 @@ public class mahasiswa04 {
     public String nim;
     public String nama;
     public String kelas;
-    public double ipk;
+    public float ipk;
 
 }
